Fix times() to set exact number of expected calls
Previusly, times() set the maximum number of calls expected. Now,
times() sets the exact number of calls expected.

This requires us to check if a Behavior is _exhausted_, which means that
it is either persistent or it has produced as often as expected. A
BehaviorStack is exhausted if all of its Behaviors is exhausted. Method
verification (calling m.verify() on a Method object) is now considered
failed if either an unexpected call was made before calling verify(), or
if state-verification is off and the BehaviorStack is not exhausted.

After adapting the tests of Method to the new behavior, all tests pass.
The tests of Behavior and BehaviorStack have yet to be changed to check
the correctness of is_exhausted.