Fix CMAKE_PREFIX_PATH in Makefiles
A previous commit erroneously assumed that using a
DCMAKE_PREFIX_PATH=... directive leads to overriding the value of the
environment variable CMAKE_PREFIX_PATH. That doesn't seem to be the
case - instead, the value of the CMake directive is appended to that of
the environment variable. The Makefiles and documentation have been
changed to reflect this.