Implement and test emit feature
This commit introduces some significant changes to the structure of
DrMock which allow mock objects to emit Qt signals when they are called.

All major class templates now have a new template parameter,
the type of the parent mock object (MO). For every MO, the mocker passes
a to each Method object of the MO a pointer to the MO. This pointer is
required for Signal::invoke(Parent*) [see below].

Behavior and StateBehavior add the `emits` method, with which the user
specifies that the behavior shall produce a signal emission, alongside
the arguments with which the signal is emitted.

A new class template, Signal, represents such signal emissions. By
calling the `invoke(Parent* parent)` method, `parent` emits the signal
with the specified arguments.

Although DrMock is designed with Qt5 in mind, non-Qt use must also be
possible. Thus, this commit introduces the QT_USE_MOCK macro. The user
really shouldn't touch this macro.

Instead, the DrMockGenerator script get's a new command line argument,
`--qt`. If `--qt` is set, then every mock object implementation is
equipped with `#define DRMOCK_USE_QT`. The DrMock CMake macros
automatically apply `--qt` if the test is linked against a Qt library.

The effect of `DRMOCK_USE_QT` is straightforward (we've made sure to
keep the impact of preprocessor macros small): If the macro is _not_
defined, then calling `Signal::invoke(Parent*)` throws an exception,
instead of emitting the specified signal.

The usage of this macro has also led to a slight reorganization of our
build system: If the environment variable `DRMOCK_QT_PATH` is set all
mocker tests run with `DRMOCK_USE_QT` defined, otherwise all tests run
without `DRMOCK_USE_QT` (except for the Qt-specific tests, which are
omitted, of course).

Emit and return behaviors may be combined (the method first emits, then
returns). Throwing may not be combined with any of them. Thus, the
`results_` field of `StateBehavior` (for example) now maps to pairs,
with first entry equal to a shared_ptr to the return value (nullptr if
not available) and second entry equal to a shared_ptr to the Signal
object (nullptr if not available).

Moreover, the refactor of `StateBehavior` has also led us to be more
clear about redefining behaviors. Calling `returns` or `emits` multiple
times on the same state throws an exception, as does calling `throws`
multiple times or combining `throws` with `returns` or `emits`.

The terminology surrounding method returns, thrown exceptions and signal
emissions is still a bit confusing. The term result sometimes refers to
a returned value or an emitted signal, but sometimes also to a thrown
exception. This should be cleaned up in the near future.