Remove DRTEST_USE_QT when using QSignalSpy
(In the case of EmitTest.cpp, there was no QSignalSpy at work, anyways.)
We may need to remove QSignalSpy entirely, due to it not being able to
sue anything other than Qt::DirectConnection (this follows from an
inspection of the source code, only). It appears as though it will
remain this way in qt6.