Fix dangeling reference
In an attempt to not overcomplicate things, we don't test direct
connection behavior with respect to copies (const references are not
copied when using direct connection, as there is no need to marshall the
data to another thread). Demonstrating this behavior would complicate
the Dummy class even further, and the correct behavior of non-const
references under direct connection has already been verified.

Also, we don't check in `invokeWithParametersQueuedConnection` how many
times the `str` is copied. It would be prudent to design a spy object
which tracks calls to the copy ctor to check that the `str` argument is
copied only once (Qt needs to copy this object when passed from one
thread to another). This will have to wait until another time.