Add Variant for implicit conversion
We use a homebrew `Variant` class which allows implicit conversion for
the first alternative in place of `std::variant` in `expect_t`. This
allows the user to use implicit conversion when calling `expects` and
`transition`, thus simplifying expressions like
`expects(std::vector<int>{1, 2, 3})` into `expects({1, 2, 3})`.

The `Variant` class is implemented as wrapper of `std::variant`.