Fix segmentation fault
This segmentation fault occurs if no return value is specified in a
behavior, forcing a `Method` to use a default constructed return value,
stored in a `shared_ptr`. But the `shared_ptr` is destroyed immediately
after dereferencing it in the mock implementation of the function. In
particular, the mock implementation copies or moves from an already
destroyed object.

Strangely enough, this was never a problem with clang, but immediately
caused a wall of segfaults on MSVC...